Move `findMockEntity` to app package and remove from lib
