+++ packages/app/src/providers/mock/mock-api.ts
@@ -1,4 +1,3 @@
-import { findMockEntity } from '@h5web/lib';
 import {
   assertArrayShape,
   assertDefined,
@@ -25,6 +24,7 @@ import { DataProviderApi } from '../api';
 import type { ExportFormat, ExportURL, ValuesStoreParams } from '../models';
 import { sliceValue } from '../utils';
 import { makeMockFile } from './mock-file';
+import { findMockEntity } from './utils';
 
 export const SLOW_TIMEOUT = 3000;
